Move coco-related logic into a dedicated file
