+++ typescript/web-app/src/connectors/resolvers/format-coco/add-image-dimensions-to-labels.ts
@@ -1,0 +1,22 @@
+import { DbLabel } from "../../database";
+import { Repository } from "../../repository/types";
+import { DbLabelWithImageDimensions } from "./coco-core/types";
+
+export const addImageDimensionsToLabels = async (
+  labels: DbLabel[],
+  repository: Repository
+): Promise<DbLabelWithImageDimensions[]> => {
+  return Promise.all(
+    labels.map(async (label) => {
+      const { imageId } = label;
+      const image = await repository.image.getById(imageId);
+      if (image == null) {
+        throw new Error(`Missing image with id ${imageId}`);
+      }
+      return {
+        ...label,
+        imageDimensions: { height: image.height, width: image.width },
+      };
+    })
+  );
+};
